sys: kern: arch: rv32: move tasks functions to own module.
