+++ sys/kern/src/arch/riscv32/mod.rs
@@ -17,7 +17,6 @@
 use core::arch::asm;
 #[cfg(feature = "custom-interrupts")]
 use core::convert::TryInto;
-use core::ptr::NonNull;
 
 use core::sync::atomic::Ordering;
 
@@ -30,15 +29,9 @@ cfg_if::cfg_if! {
     }
 }
 
-use crate::task;
 use crate::time::Timestamp;
-use crate::umem::USlice;
-use unwrap_lite::UnwrapLite;
 
 extern crate riscv_rt;
-
-use riscv::register;
-use riscv::register::mstatus::MPP;
 
 #[allow(unused)]
 macro_rules! uassert {
@@ -61,10 +54,8 @@ pub use trap::*;
 mod saved_state;
 pub use saved_state::*;
 
-/// On RISC-V we use a global to record the current task pointer.  It may be
-/// possible to use the mscratch register instead.
-#[no_mangle]
-pub static mut CURRENT_TASK_PTR: Option<NonNull<task::Task>> = None;
+mod task;
+pub use task::*;
 
 /// To allow our clock frequency to be easily determined from a debugger, we
 /// store it in memory.
@@ -83,39 +74,6 @@ pub fn set_clock_freq(tick_divisor: u32) {
     unsafe {
         CLOCK_FREQ_KHZ = tick_divisor;
     }
-}
-
-pub fn reinitialize(task: &mut task::Task) {
-    *task.save_mut() = SavedState::default();
-
-    // Set the initial stack pointer, ensuring 16-byte stack alignment as per
-    // the RISC-V calling convention.
-    let initial_stack = task.descriptor().initial_stack;
-    task.save_mut().set_sp(initial_stack);
-    uassert!(task.save().sp() & 0xF == 0);
-
-    // zap the stack with a distinct pattern
-    for region in task.region_table().iter() {
-        if initial_stack < region.base {
-            continue;
-        }
-        if initial_stack > region.base + region.size {
-            continue;
-        }
-        let mut uslice: USlice<u32> = USlice::from_raw(
-            region.base as usize,
-            (initial_stack as usize - region.base as usize) >> 2,
-        )
-        .unwrap_lite();
-
-        let zap = task.try_write(&mut uslice).unwrap_lite();
-        for word in zap.iter_mut() {
-            *word = 0xbaddcafe;
-        }
-    }
-    // Set the initial program counter
-    let pc = task.descriptor().entry_point;
-    task.save_mut().set_pc(pc);
 }
 
 cfg_if::cfg_if! {
@@ -160,45 +118,6 @@ cfg_if::cfg_if! {
     }
 }
 
-#[allow(unused_variables)]
-pub fn start_first_task(tick_divisor: u32, task: &mut task::Task) -> ! {
-    unsafe {
-        //
-        // Configure the timer
-        //
-        CLOCK_FREQ_KHZ = tick_divisor;
-
-        // make MTIMECMP start with MTIME
-        let mtime = core::ptr::read_volatile(crate::startup::MTIME as *mut u64);
-        core::ptr::write_volatile(crate::startup::MTIMECMP as *mut u64, mtime);
-
-        // increment mtimecmp for appropriate timer interrupts
-        reset_timer();
-
-        // Machine timer interrupt enable
-        register::mie::set_mtimer();
-
-        // Configure MPP to switch us to User mode on exit from Machine
-        // mode (when we call "mret" below).
-        register::mstatus::set_mpp(MPP::User);
-    }
-
-    // Write the initial task program counter.
-    register::mepc::write(task.save().pc() as *const usize as usize);
-
-    // Load first task pointer, set its initial stack pointer, and exit out
-    // of machine mode, launching the task.
-    unsafe {
-        crate::task::activate_next_task(task);
-        asm!("
-            lw sp, ({sp})
-            mret",
-            sp = in(reg) &task.save().sp(),
-            options(noreturn)
-        );
-    }
-}
-
 impl crate::atomic::AtomicExt for AtomicBool {
     type Primitive = bool;
 
@@ -209,21 +128,6 @@ impl crate::atomic::AtomicExt for AtomicBool {
         ordering: Ordering,
     ) -> Self::Primitive {
         self.swap(value, ordering)
-    }
-}
-
-/// Records the address of `task` as the current user task.
-///
-/// # Safety
-///
-/// This records a pointer that aliases `task`. As long as you don't read that
-/// pointer while you have access to `task`, and as long as the `task` being
-/// stored is actually in the taask table, you'll be okay.
-pub unsafe fn set_current_task(task: &mut task::Task) {
-    // Safety: should be ok if the contract above is met
-    // TODO: make me an atomic
-    unsafe {
-        CURRENT_TASK_PTR = Some(NonNull::from(task));
     }
 }
